refactor(domain): move error codes to another file
